+++ taotao-cloud-microservice/taotao-cloud-distribution/taotao-cloud-distribution-biz/src/main/java/com/taotao/cloud/distribution/biz/service/impl/DistributionCashServiceImpl.java
@@ -44,7 +44,7 @@ public class DistributionCashServiceImpl extends ServiceImpl<DistributionCashMap
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public Boolean cash(Double applyMoney) {
+    public Boolean cash(BigDecimal applyMoney) {
 
         //检查分销功能开关
         distributionService.checkDistributionSetting();
