+++ taotao-cloud-microservice/taotao-cloud-order/taotao-cloud-order-biz/src/main/java/com/taotao/cloud/order/biz/controller/seller/AfterSaleController.java
@@ -73,7 +73,7 @@ public class AfterSaleController {
 		@NotNull(message = "请选择售后单") @PathVariable String afterSaleSn,
 		@NotNull(message = "请审核") String serviceStatus,
 		String remark,
-		Double actualRefundPrice) {
+		BigDecimal actualRefundPrice) {
 
 		return Result.success(
 			afterSaleService.review(afterSaleSn, serviceStatus, remark, actualRefundPrice));
